Adapted to api changes

+++ src/test/java/no/ssb/rawdata/pulsar/PulsarRawdataClientTck.java
@@ -3,8 +3,8 @@ package no.ssb.rawdata.pulsar;
 import no.ssb.rawdata.api.RawdataClient;
 import no.ssb.rawdata.api.RawdataClientInitializer;
 import no.ssb.rawdata.api.RawdataConsumer;
-import no.ssb.rawdata.api.RawdataContentNotBufferedException;
 import no.ssb.rawdata.api.RawdataMessage;
+import no.ssb.rawdata.api.RawdataNotBufferedException;
 import no.ssb.rawdata.api.RawdataProducer;
 import no.ssb.service.provider.api.ProviderConfigurator;
 import org.apache.pulsar.client.admin.Namespaces;
@@ -60,15 +60,20 @@ public class PulsarRawdataClientTck {
             Topics topics = admin.topics();
 
             if (tenants.getTenants().contains(tenant)) {
-                namespaces.getNamespaces(tenant).contains(namespace);
+                if (namespaces.getNamespaces(tenant).contains(namespace)) {
 
-                // delete all topics in namespace
-                List<String> existingTopics = namespaces.getTopics(namespace);
-                for (String topic : existingTopics) {
-                    topics.delete(topic);
+                    // delete all topics in namespace
+                    List<String> existingTopics = namespaces.getTopics(namespace);
+                    for (String topic : existingTopics) {
+                        List<String> subscriptions = topics.getSubscriptions(topic);
+                        for (String subscription : subscriptions) {
+                            topics.deleteSubscription(topic, subscription); // TODO doesn't work when there are active producers/consumers
+                        }
+                        topics.delete(topic);
+                    }
+
+                    namespaces.deleteNamespace(namespace);
                 }
-
-                namespaces.deleteNamespace(namespace);
 
                 tenants.deleteTenant(tenant);
             }
@@ -86,107 +91,93 @@ public class PulsarRawdataClientTck {
     }
 
     @Test
-    public void thatLastPositionOfEmptyTopicCanBeReadByProducer() throws Exception {
-        try (RawdataProducer producer = client.producer("the-topic")) {
-
-            assertEquals(producer.lastPosition(), null);
-        }
+    public void thatLastPositionOfEmptyTopicCanBeRead() {
+        assertEquals(client.lastPosition("the-topic"), null);
     }
 
     @Test
-    public void thatLastPositionOfProducerCanBeRead() throws Exception {
-        try (RawdataProducer producer = client.producer("the-topic")) {
+    public void thatLastPositionOfProducerCanBeRead() {
+        RawdataProducer producer = client.producer("the-topic");
 
-            producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
-            producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
-            producer.publish("a", "b");
+        producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+        producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
+        producer.publish("a", "b");
 
-            assertEquals(producer.lastPosition(), "b");
+        assertEquals(client.lastPosition("the-topic"), "b");
 
-            producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
-            producer.publish("c");
+        producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
+        producer.publish("c");
 
-            assertEquals(producer.lastPosition(), "c");
-        }
+        assertEquals(client.lastPosition("the-topic"), "c");
     }
 
-    @Test(expectedExceptions = RawdataContentNotBufferedException.class)
-    public void thatPublishNonBufferedMessagesThrowsException() throws Exception {
-        try (RawdataProducer producer = client.producer("the-topic")) {
-            producer.publish("unbuffered-1");
-        }
-    }
-
-    @Test
-    public void thatSingleMessageCanBeProducedAndConsumerSynchronously() throws Exception {
-        RawdataMessage expected;
-        try (RawdataProducer producer = client.producer("the-topic")) {
-            expected = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
-            producer.publish(expected.position());
-        }
-        try (RawdataConsumer consumer = client.consumer("the-topic")) {
-            RawdataMessage message = consumer.receive(1, TimeUnit.SECONDS);
-            assertEquals(message, expected);
-        }
+    @Test(expectedExceptions = RawdataNotBufferedException.class)
+    public void thatPublishNonBufferedMessagesThrowsException() {
+        RawdataProducer producer = client.producer("the-topic");
+        producer.publish("unbuffered-1");
     }
 
     @Test
-    public void thatSingleMessageCanBeProducedAndConsumerAsynchronously() throws Exception {
-        try (RawdataConsumer consumer = client.consumer("the-topic")) {
-            CompletableFuture<? extends RawdataMessage> future = consumer.receiveAsync();
-            RawdataMessage expected;
-            try (RawdataProducer producer = client.producer("the-topic")) {
-                expected = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
-                producer.publish(expected.position());
-            }
-            RawdataMessage message = future.join();
-            assertEquals(message, expected);
-        }
+    public void thatSingleMessageCanBeProducedAndConsumerSynchronously() throws InterruptedException {
+        RawdataProducer producer = client.producer("the-topic");
+        RawdataConsumer consumer = client.consumer("the-topic");
+
+        RawdataMessage expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+        producer.publish(expected1.position());
+
+        RawdataMessage message = consumer.receive(1, TimeUnit.SECONDS);
+        assertEquals(message, expected1);
     }
 
     @Test
-    public void thatMultipleMessagesCanBeProducedAndConsumerSynchronously() throws Exception {
-        RawdataMessage expected1;
-        RawdataMessage expected2;
-        RawdataMessage expected3;
-        try (RawdataProducer producer = client.producer("the-topic")) {
-            expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
-            expected2 = producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
-            expected3 = producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
-            producer.publish(expected1.position(), expected2.position(), expected3.position());
-        }
-        try (RawdataConsumer consumer = client.consumer("the-topic")) {
-            RawdataMessage message1 = consumer.receive(1, TimeUnit.SECONDS);
-            RawdataMessage message2 = consumer.receive(1, TimeUnit.SECONDS);
-            RawdataMessage message3 = consumer.receive(1, TimeUnit.SECONDS);
-            assertEquals(message1, expected1);
-            assertEquals(message2, expected2);
-            assertEquals(message3, expected3);
-        }
+    public void thatSingleMessageCanBeProducedAndConsumerAsynchronously() {
+        RawdataProducer producer = client.producer("the-topic");
+        RawdataConsumer consumer = client.consumer("the-topic");
+
+        CompletableFuture<? extends RawdataMessage> future = consumer.receiveAsync();
+
+        RawdataMessage expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+        producer.publish(expected1.position());
+
+        RawdataMessage message = future.join();
+        assertEquals(message, expected1);
     }
 
     @Test
-    public void thatMultipleMessagesCanBeProducedAndConsumerAsynchronously() throws Exception {
-        try (RawdataConsumer consumer = client.consumer("the-topic")) {
+    public void thatMultipleMessagesCanBeProducedAndConsumerSynchronously() throws InterruptedException {
+        RawdataProducer producer = client.producer("the-topic");
+        RawdataConsumer consumer = client.consumer("the-topic");
 
-            CompletableFuture<List<RawdataMessage>> future = receiveAsyncAddMessageAndRepeatRecursive(consumer, "c", new ArrayList<>());
+        RawdataMessage expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+        RawdataMessage expected2 = producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
+        RawdataMessage expected3 = producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
+        producer.publish(expected1.position(), expected2.position(), expected3.position());
 
-            RawdataMessage expected1;
-            RawdataMessage expected2;
-            RawdataMessage expected3;
-            try (RawdataProducer producer = client.producer("the-topic")) {
-                expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
-                expected2 = producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
-                expected3 = producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
-                producer.publish(expected1.position(), expected2.position(), expected3.position());
-            }
+        RawdataMessage message1 = consumer.receive(1, TimeUnit.SECONDS);
+        RawdataMessage message2 = consumer.receive(1, TimeUnit.SECONDS);
+        RawdataMessage message3 = consumer.receive(1, TimeUnit.SECONDS);
+        assertEquals(message1, expected1);
+        assertEquals(message2, expected2);
+        assertEquals(message3, expected3);
+    }
 
-            List<RawdataMessage> messages = future.join();
+    @Test
+    public void thatMultipleMessagesCanBeProducedAndConsumerAsynchronously() {
+        RawdataProducer producer = client.producer("the-topic");
+        RawdataConsumer consumer = client.consumer("the-topic");
 
-            assertEquals(messages.get(0), expected1);
-            assertEquals(messages.get(1), expected2);
-            assertEquals(messages.get(2), expected3);
-        }
+        CompletableFuture<List<RawdataMessage>> future = receiveAsyncAddMessageAndRepeatRecursive(consumer, "c", new ArrayList<>());
+
+        RawdataMessage expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+        RawdataMessage expected2 = producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
+        RawdataMessage expected3 = producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
+        producer.publish(expected1.position(), expected2.position(), expected3.position());
+
+        List<RawdataMessage> messages = future.join();
+
+        assertEquals(messages.get(0), expected1);
+        assertEquals(messages.get(1), expected2);
+        assertEquals(messages.get(2), expected3);
     }
 
     private CompletableFuture<List<RawdataMessage>> receiveAsyncAddMessageAndRepeatRecursive(RawdataConsumer consumer, String endPosition, List<RawdataMessage> messages) {
@@ -200,37 +191,28 @@ public class PulsarRawdataClientTck {
     }
 
     @Test
-    public void thatMessagesCanBeConsumedByMultipleConsumers() throws Exception {
-        try (RawdataConsumer consumer1 = client.consumer("the-topic")) {
+    public void thatMessagesCanBeConsumedByMultipleConsumers() {
+        RawdataProducer producer = client.producer("the-topic");
+        RawdataConsumer consumer1 = client.consumer("the-topic");
+        RawdataConsumer consumer2 = client.consumer("the-topic");
 
-            RawdataMessage expected1;
-            RawdataMessage expected2;
-            RawdataMessage expected3;
+        CompletableFuture<List<RawdataMessage>> future1 = receiveAsyncAddMessageAndRepeatRecursive(consumer1, "c", new ArrayList<>());
+        CompletableFuture<List<RawdataMessage>> future2 = receiveAsyncAddMessageAndRepeatRecursive(consumer2, "c", new ArrayList<>());
 
-            CompletableFuture<List<RawdataMessage>> future1 = receiveAsyncAddMessageAndRepeatRecursive(consumer1, "c", new ArrayList<>());
-            try (RawdataConsumer consumer2 = client.consumer("the-topic")) {
+        RawdataMessage expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+        RawdataMessage expected2 = producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
+        RawdataMessage expected3 = producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
+        producer.publish(expected1.position(), expected2.position(), expected3.position());
 
-                CompletableFuture<List<RawdataMessage>> future2 = receiveAsyncAddMessageAndRepeatRecursive(consumer2, "c", new ArrayList<>());
+        List<RawdataMessage> messages1 = future1.join();
+        assertEquals(messages1.get(0), expected1);
+        assertEquals(messages1.get(1), expected2);
+        assertEquals(messages1.get(2), expected3);
 
-                try (RawdataProducer producer = client.producer("the-topic")) {
-                    expected1 = producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
-                    expected2 = producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
-                    expected3 = producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
-                    producer.publish(expected1.position(), expected2.position(), expected3.position());
-                }
-
-
-                List<RawdataMessage> messages2 = future2.join();
-                assertEquals(messages2.get(0), expected1);
-                assertEquals(messages2.get(1), expected2);
-                assertEquals(messages2.get(2), expected3);
-            }
-
-            List<RawdataMessage> messages1 = future1.join();
-            assertEquals(messages1.get(0), expected1);
-            assertEquals(messages1.get(1), expected2);
-            assertEquals(messages1.get(2), expected3);
-        }
+        List<RawdataMessage> messages2 = future2.join();
+        assertEquals(messages2.get(0), expected1);
+        assertEquals(messages2.get(1), expected2);
+        assertEquals(messages2.get(2), expected3);
     }
 
     @Test
@@ -305,6 +287,46 @@ public class PulsarRawdataClientTck {
         try (RawdataConsumer consumer = client.consumer("the-topic", "d")) {
             RawdataMessage message = consumer.receive(100, TimeUnit.MILLISECONDS);
             assertNull(message);
+        }
+    }
+
+    @Test
+    public void thatSeekToWorks() throws Exception {
+        long timestampBeforeA;
+        long timestampBeforeB;
+        long timestampBeforeC;
+        long timestampBeforeD;
+        long timestampAfterD;
+        try (RawdataProducer producer = client.producer("the-topic")) {
+            producer.buffer(producer.builder().position("a").put("payload", new byte[5]));
+            timestampBeforeA = System.currentTimeMillis();
+            producer.publish("a");
+            Thread.sleep(5);
+            producer.buffer(producer.builder().position("b").put("payload", new byte[3]));
+            timestampBeforeB = System.currentTimeMillis();
+            producer.publish("b");
+            Thread.sleep(5);
+            producer.buffer(producer.builder().position("c").put("payload", new byte[7]));
+            timestampBeforeC = System.currentTimeMillis();
+            producer.publish("c");
+            Thread.sleep(5);
+            producer.buffer(producer.builder().position("d").put("payload", new byte[7]));
+            timestampBeforeD = System.currentTimeMillis();
+            producer.publish("d");
+            Thread.sleep(5);
+            timestampAfterD = System.currentTimeMillis();
+        }
+        try (RawdataConsumer consumer = client.consumer("the-topic")) {
+            consumer.seek(timestampAfterD);
+            assertNull(consumer.receive(100, TimeUnit.MILLISECONDS));
+            consumer.seek(timestampBeforeD);
+            assertEquals("d", consumer.receive(100, TimeUnit.MILLISECONDS).position());
+            consumer.seek(timestampBeforeB);
+            assertEquals("b", consumer.receive(100, TimeUnit.MILLISECONDS).position());
+            consumer.seek(timestampBeforeC);
+            assertEquals("c", consumer.receive(100, TimeUnit.MILLISECONDS).position());
+            consumer.seek(timestampBeforeA);
+            assertEquals("a", consumer.receive(100, TimeUnit.MILLISECONDS).position());
         }
     }
 }
